requirements.txt
flask
inference_sdk
pillow

